client: Move templates closer to the source
React projects typically do not distinguish between templates and controlling code,
instead they make use of self-contained components.
