+++ assets/js/selfoss-ui.js
@@ -614,9 +614,12 @@ selfoss.ui = {
             {label}
         </button>);
 
-        messageContainer.html([
-            message
-        ].concat(buttons));
+        messageContainer.html(
+            <React.Fragment>
+                {message}
+                {buttons}
+            </React.Fragment>
+        );
 
         if (isError) {
             messageContainer.addClass('error');
